feat: handle ctrl+c by key reader
